Move test models from attr package to upper package
