+++ tests/attr/layer/test_internal_influence.py
@@ -10,7 +10,7 @@ from captum._utils.typing import BaselineType
 from captum.attr._core.layer.internal_influence import InternalInfluence
 
 from ...helpers.basic import BaseTest, assertTensorTuplesAlmostEqual
-from ..helpers.basic_models import (
+from ...helpers.basic_models import (
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
 )
