+++ tests/attr/layer/test_layer_activation.py
@@ -15,7 +15,7 @@ from ...helpers.basic import (
     assertTensorAlmostEqual,
     assertTensorTuplesAlmostEqual,
 )
-from ..helpers.basic_models import (
+from ...helpers.basic_models import (
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
 )
