+++ tests/attr/layer/test_layer_deeplift_basic.py
@@ -16,7 +16,7 @@ from ...helpers.basic import (
     assertTensorAlmostEqual,
     assertTensorTuplesAlmostEqual,
 )
-from ..helpers.basic_models import (
+from ...helpers.basic_models import (
     BasicModel_MultiLayer,
     LinearMaxPoolLinearModel,
     ReLULinearDeepLiftModel,
