+++ tests/attr/neuron/test_neuron_integrated_gradients.py
@@ -18,7 +18,7 @@ from ...helpers.basic import (
     assertArraysAlmostEqual,
     assertTensorTuplesAlmostEqual,
 )
-from ..helpers.basic_models import (
+from ...helpers.basic_models import (
     BasicModel_ConvNet,
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
