+++ tests/attr/test_data_parallel.py
@@ -32,7 +32,7 @@ from captum.attr._core.occlusion import Occlusion
 from captum.attr._core.shapley_value import ShapleyValues, ShapleyValueSampling
 
 from ..helpers.basic import BaseGPUTest
-from .helpers.basic_models import (
+from ..helpers.basic_models import (
     BasicModel_ConvNet,
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
