+++ tests/attr/test_deconvolution.py
@@ -15,7 +15,7 @@ from captum.attr._core.neuron.neuron_guided_backprop_deconvnet import (
 )
 
 from ..helpers.basic import BaseTest, assertTensorAlmostEqual
-from .helpers.basic_models import BasicModel_ConvNet_One_Conv
+from ..helpers.basic_models import BasicModel_ConvNet_One_Conv
 
 
 class Test(BaseTest):
