+++ tests/attr/test_feature_permutation.py
@@ -7,7 +7,7 @@ from torch import Tensor
 from captum.attr._core.feature_permutation import FeaturePermutation, _permute_feature
 
 from ..helpers.basic import BaseTest, assertArraysAlmostEqual, assertTensorAlmostEqual
-from .helpers.basic_models import BasicModelWithSparseInputs
+from ..helpers.basic_models import BasicModelWithSparseInputs
 
 
 class Test(BaseTest):
