+++ tests/attr/test_guided_grad_cam.py
@@ -10,7 +10,7 @@ from captum._utils.typing import TensorOrTupleOfTensorsGeneric
 from captum.attr._core.guided_grad_cam import GuidedGradCam
 
 from ..helpers.basic import BaseTest, assertTensorAlmostEqual
-from .helpers.basic_models import BasicModel_ConvNet_One_Conv
+from ..helpers.basic_models import BasicModel_ConvNet_One_Conv
 
 
 class Test(BaseTest):
