+++ tests/attr/test_shapley.py
@@ -9,7 +9,7 @@ from captum._utils.typing import BaselineType, TensorOrTupleOfTensorsGeneric
 from captum.attr._core.shapley_value import ShapleyValues, ShapleyValueSampling
 
 from ..helpers.basic import BaseTest, assertTensorTuplesAlmostEqual
-from .helpers.basic_models import (
+from ..helpers.basic_models import (
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
 )
